Add a Color enumeration
This enumeration represents the two colors of the pieces on the board,
it should allow you to refer to a color as `Color.BLUE' or `Color.RED'
as opposed to having to use strings or random numbers or whatever. It
also means that if for whatever reason we decide to flip the positions
of blue and red or something similiar, it should be a very quick and
easy fix.