Add a game script
At the moment this script does not do much, it simply registers an
event listener to inform the server about when one of the clients
resigns the game which activates when someone clicks the "resign"
button. In the future this script will most likely house all the
client-side functionality related to the game.