Create a game class
The game class contains the entire state of the game, its history, and
the websockets connected to both clients. The class also has a
`messageAll()' method which can be used to send a message to both of
the clients. The `boardInit()' function can be used to initialize a
brand new 10x10 draughts board.