Add a game queue and handle disconnects
The server now has an array of games that currently exist. When a
client attempts to join a game, the server will check if another
player is already waiting for an opponent and if so, matches the two
players against each other. If all the games are full the server will
create a new game and the player will need to wait for an opponent.

The server also has handling for clients that disconnect. When that
event occurs, the remaining client is sent a message informing them
that their opponent disconnected, and the game is deleted. None of the
statistics are updated as it wouldn't make sense to count a statistic
such as `minimumMoves' if the opponent disconnected only a few moves in.